sign_to_shop/requirements.txt
click==8.1.7
colorama==0.4.6
csvtotable==2.1.1
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
unicodecsv==0.14.1
